Controller fonctionnel avec tous les endpoints

+++ src/main/java/com/openclassrooms/p9/microservice_patient/controller/PatientController.java
@@ -2,13 +2,10 @@ package com.openclassrooms.p9.microservice_patient.controller;
 
 import com.openclassrooms.p9.microservice_patient.model.Patient;
 import com.openclassrooms.p9.microservice_patient.service.PatientService;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -23,40 +20,79 @@ public class PatientController {
     @Autowired
     PatientService patientService;
 
-    @RequestMapping("/Patients/list")
+    @GetMapping("/Patients")
     public List<Patient> listAll()
     {
-        List<Patient> p=patientService.getAll();
-        System.out.println(p.get(2).getPrenom());
         return patientService.getAll();
     }
 
     @GetMapping(value = "/Patients/{id}")
-    public Optional<Patient> showPatient(@PathVariable int id) {
-        return patientService.getPatientById(id);
+    public Patient showPatient(@PathVariable int id) {
+        Optional<Patient> patientFound = patientService.getPatientById(id);
+        if (patientFound.isPresent()) {
+             return patientFound.get();
+        }
+        else{
+            throw new PatientNotFoundException();
+        }
     }
 
+    @PutMapping(value = "/Patients/{id}")
+    public ResponseEntity<Patient> updatePatient(@Valid @RequestBody Patient patient,@PathVariable int id) {
+        Optional<Patient> patientFound = patientService.getPatientById(id);
+        if (patientFound.isPresent()) {
+            patient.setId(id);
+            Patient patientUpdated = patientService.savePatient(patient);
+
+            if (Objects.isNull(patientUpdated)) {
+                return ResponseEntity.noContent().build();
+            }
+
+            URI location = ServletUriComponentsBuilder
+                    .fromCurrentRequest()
+                    .path("/{id}")
+                    .buildAndExpand(patientUpdated.getId())
+                    .toUri();
+            return ResponseEntity.ok(patientUpdated);
+        }
+        else{
+            throw new PatientNotFoundException();
+        }
+    }
 
     @PostMapping(value = "/Patients")
-    public ResponseEntity<Patient> addPatient(@RequestBody Patient patient) {
+    public ResponseEntity<Patient> addPatient(@Valid @RequestBody Patient patient) {
         Patient patientAdded = patientService.savePatient(patient);
-        //System.out.println(patient.getId());
+
         if (Objects.isNull(patientAdded)) {
             return ResponseEntity.noContent().build();
         }
-        //.buildAndExpand(patientAdded.getId())
+
         URI location = ServletUriComponentsBuilder
                 .fromCurrentRequest()
                 .path("/{id}")
-                .buildAndExpand(2)
+                .buildAndExpand(patientAdded.getId())
                 .toUri();
         return ResponseEntity.created(location).build();
     }
 
 
-    @GetMapping("/Patients/delete/{id}")
-    public String deletePatient(@PathVariable("id") Integer id) {
-        patientService.deletePatientById(id);
-        return "redirect:/Patients/list";
+    @DeleteMapping ("/Patients/{id}")
+    public void deletePatient(@PathVariable("id") Integer id) {
+        Optional<Patient> patientFound = patientService.getPatientById(id);
+        if (patientFound.isPresent()) {
+            patientService.deletePatientById(id);
+        }
+        else {
+            throw new PatientNotFoundException();
+        }
+    }
+
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public static class PatientNotFoundException extends RuntimeException {
+        public PatientNotFoundException() {
+            super("Patient not found");
+        }
     }
 }
